requirements.txt
Flask
Flask-SqlAlchemy
Flask-Script
six
config
Flask-Restful
Flask-Migrate